frontend: add route page
