+++ frontend/src/App.css
@@ -36,3 +36,7 @@
     transform: rotate(360deg);
   }
 }
+
+header, footer {
+  height: 10%;
+}
